+++ compiler/src/iree/compiler/Dialect/Flow/Transforms/FormDispatchRegions.cpp
@@ -24,10 +24,8 @@
 #include "mlir/Dialect/Linalg/IR/Linalg.h"
 #include "mlir/Dialect/Linalg/Transforms/Transforms.h"
 #include "mlir/Dialect/Linalg/Utils/Utils.h"
-#include "mlir/Dialect/MemRef/Transforms/Passes.h"
 #include "mlir/Dialect/SCF/IR/SCF.h"
 #include "mlir/Dialect/Tensor/IR/Tensor.h"
-#include "mlir/Dialect/Tensor/Transforms/Transforms.h"
 #include "mlir/Dialect/Utils/StructuredOpsUtils.h"
 #include "mlir/IR/Block.h"
 #include "mlir/IR/Builders.h"
@@ -41,7 +39,6 @@
 #include "mlir/Interfaces/TilingInterface.h"
 #include "mlir/Pass/Pass.h"
 #include "mlir/Support/LLVM.h"
-#include "mlir/Transforms/GreedyPatternRewriteDriver.h"
 #include "mlir/Transforms/TopologicalSortUtils.h"
 
 #define DEBUG_TYPE "iree-flow-form-dispatch-regions"
@@ -848,16 +845,6 @@ void FormDispatchRegionsPass::runOnOperation() {
   if (failed(createFusionGroups(rewriter, funcOp, dominanceInfo,
                                 generateWorkloadRegion, aggressiveFusion))) {
     funcOp->emitOpError("failed to create fusion groups");
-    return signalPassFailure();
-  }
-
-  RewritePatternSet canonicalizationPatterns(&getContext());
-  memref::populateResolveRankedShapeTypeResultDimsPatterns(
-      canonicalizationPatterns);
-  tensor::populateFoldTensorEmptyPatterns(canonicalizationPatterns);
-  if (failed(applyPatternsAndFoldGreedily(
-          funcOp, std::move(canonicalizationPatterns)))) {
-    funcOp->emitOpError("failed to apply cleanup patterns");
     return signalPassFailure();
   }
 }
